Renamed files in Components folder
